Downloads/ai-internship-matcher/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
pdfminer.six
python-docx
sentence-transformers
torch
transformers
numpy
pandas
beautifulsoup4
requests
spacy
scikit-learn
joblib
openai
llama-cpp-python
python-dotenv
